Renames SubsectionDescription to AccordionLayout
